Модуляция. правки по замечаниям
