--- js/created-id-generator.js
@@ -1,10 +1,0 @@
-const createdIdGenerator = () => {
-  let i = 0;
-
-  return function () {
-    i += 1;
-    return i;
-  };
-};
-
-export {createdIdGenerator};
